B-02018: Admin 2.0: System/Java Properties: Create the Ext model and mock json data :
Use the accounts as a template.

Follow the MVC architecture used for Users.

Study and follow the file structure used in users.

The following fields are required in this round

- name (the Property name)
- value (the Property value)

In the Ext model, all the fields should be String.

Add records to the mock json file

It is not neccessary to mock all Java properties for now. Add 15 - 20 properties as an example( see System/Java Properties in 4.6)