B-02010: Admin 2.0: Contenttypes: Create grid list view
Create a grid that lists the mock data

The Grid should have a toolbar with a regular button "New" at the top. The action for the button should just alert "Not implemented" for now.

Use Accounts as an example.

The table should have the following columns:

- Name (default sort)
- Content type no
- Modified

When an item is selected in the grid the details view should be updated (see backlog item for details view).

Follow the MVC architecture used for Users.