B-02125: Admin 2.0: Accounts: Edit User Tab Panel: Member of ...
NOTE! This task also serves as a start for prototyping of CMS pickers.

In this round we only want to have a very basic group picker (no account/group list in the picker window etc.etc.). Study how the Member of... panel works in 4.6 to get an overview.

Task:

The member of panel should look and behave similar to how the member of panel behave today

    An add button (opens a group picker)
        When the add button is clicked a Group Picker/Window should be shown (in the future the window will contain the accounts/group list)
    A group membership table
        For now build the table based on static data
        Each row should contain a hidden input field with the group key and the group qualified name should be displayed together with a remove button.
        The remove button should remove the group membership (both visually and the hidden groupKey field)from the view.
    In the Picker Window create a simple button/link (later this will contain a Accounts/group list). When the button is clicked a new row should be added to the member of table (see bullet 2).