B-02117: Admin 2.0: Accounts: Ajax user delete
When delete is pressed, pass down the aciton by calling the ajax service /admin/data/user/delete. This needs to be implemented both on client and server. Use POST (not DELETE) http method. No need to actually delete the user in backend, just log out a message on the server so we can see that all is ok. Implement on the UsersResource that is already in place.

Update doc @ http://wiki.enonic.com/display/dev/Enonic+5.0+Admin+Ajax+Api after committed.