B-02192: Admin 2.0: Accounts: Layout changes for Edit User Form
Update the edit user form with the new layout specified in this wiki page:
http://wiki.enonic.com/display/dev/Task+-+Layout+changes+for+Edit+User+Form

There is an ext-js designer project file attached in the page that may be helpful to see the structure of the form.
This form will be used only for editing, not for creating a new user.
The form field-sets on the left under the toolbar (Save, cancel...) should be on top of a panel that can scroll vertically to allow for showing all the possible user fields configured. This excludes the accordion panel on the right, that should not scroll.

All the field validation and logic should still work as it does with the current form. Except the following:
- the password and repeat password fields should be removed, since this will be done with a separate action.
- The username field is moved from the User group, and will be displayed on the top header (grey background)

The buttons in the toolbar should have text and icon.

There is a separate task to implement the Address blocks in the new form layout.