B-02144: Admin 2.0: Accounts: User Group panel
Implement the User Group panel on the Edit user form accordion on the right.
See wiki page for layout and details: http://wiki.enonic.com/display/dev/Task+-+User+Group+panel

The functionality should be basically as it currently is in the Member of.. tab.
- Use an autocomplete dropdown for selecting groups
- Show each group as "buttons" with red remove icon to the right
- In order to remove groups, press red icon or drag "button" outside panel.
- when anything is changed in groups (new added/things removed) a "save changes button" must appear which the user must press to save the changes.

This task should be implemented after B-02141 is completed.