B-02246: Admin 2.0: Add functionality to edit Header and selected item box
We have started using XTemplates for parts of admin 2.0 and we need to add functionality to the elements in the templates.

1. Multiselect boxes

http://wiki.enonic.com/display/dev/*Task+-+Restructure+Acount+multiselect+view

When clicking the remove icon (red X) the item should be removed from the selection.

2. Edit form header

http://wiki.enonic.com/display/dev/Task+-+Replace+Edit+User+Form+Header

When toggling the pencil icon to the right of the User display name the display name should be editable/readonly.
When mode is editable it should be possible to edit the display name field.
Default is readonly (like it is now).