B-02142: Admin 2.0: Accounts: User Form Address grouping
Update the address groups in the edit user form according to the layout and description in this wiki page:
http://wiki.enonic.com/display/dev/Task+-+User+Form+Address+grouping

New address blocks should be added with the New Address button. They should be added inline above the button, instead of with tabs as it was before.
It should be possible to re-order the addresses with drag&drop, in a similar way as the dashlets work. Drag and drop should work only vertically within the Address panel.
The address "dashlet windows" should only have the close button on the top right (X) to allow for removing address blocks.
This task should be implemented after B-02141 is completed.

A Wiki document should be created!